// src/app/models/rastreabilidade.ts
export class Rastreabilidade {

    codigo: string;
    usuario: string='';
   // data_posicao: string;
    latitude: string;
    longitude: string;
    pim: string;
    imei: string;
    versao: string;
    Datagrup: string;
    horaInicial:string='';
    horaFinal:string='';
    data_posicao2: string='';
    data_insercao: string;
    grupo:string='';
    segmento:string ='';
    dataInicial: string = '';
    dataFinal: string = '';
    customMarker: string = '';

}